mv MessageLayer used by Roost + Wren to aviary.core, add better MessageLayer repr too
